Implmented pip install, precomputation and added suggestions
