+++ .gitignore
@@ -1,1 +1,5 @@
 __pycache__/
+.vscode
+.venv
+resources/.parquet
+.pytest_cache
