+++ requirements.txt
@@ -1,0 +1,23 @@
+attrs==21.4.0
+cycler==0.11.0
+fonttools==4.29.1
+iniconfig==1.1.1
+kiwisolver==1.3.2
+llvmlite==0.38.0
+matplotlib==3.5.1
+numba==0.55.1
+numpy==1.21.5
+packaging==21.3
+pandas==1.4.0
+Pillow==9.0.0
+pluggy==1.0.0
+py==1.11.0
+pyarrow==6.0.1
+pyparsing==3.0.7
+pytest==6.2.5
+python-dateutil==2.8.2
+pytz==2021.3
+scipy==1.7.3
+seaborn==0.11.2
+six==1.16.0
+toml==0.10.2
